+++ src/handlers/onramp/properties.rs
@@ -17,10 +17,13 @@ use {
 pub struct QueryParams {
     pub r#type: PropertyType,
     pub project_id: String,
+    /// Comma separated list of countries to filter by
     pub countries: Option<String>,
+    /// Comma separated list of provider names to exclude
+    pub exclude_providers: Option<String>,
 }
 
-#[derive(Debug, Serialize, Deserialize, Clone)]
+#[derive(Debug, Serialize, Deserialize, Clone, PartialEq)]
 #[serde(rename_all = "kebab-case")]
 pub enum PropertyType {
     Countries,