+++ src/handlers/wallet/exchanges/mod.rs
@@ -1,5 +1,8 @@
 use {
-    crate::{state::AppState, utils::crypto::Caip19Asset},
+    crate::{
+        state::AppState,
+        utils::{crypto, crypto::Caip19Asset},
+    },
     axum::extract::State,
     serde::{Deserialize, Serialize},
     std::sync::Arc,
@@ -11,9 +14,11 @@ use {
 
 pub mod binance;
 pub mod coinbase;
+pub mod test_exchange;
 
 use binance::BinanceExchange;
 use coinbase::CoinbaseExchange;
+use test_exchange::TestExchange;
 
 #[derive(Debug, Clone, Deserialize, Eq, PartialEq)]
 pub struct Config {
@@ -82,6 +87,7 @@ pub trait ExchangeProvider {
 pub enum ExchangeType {
     Binance,
     Coinbase,
+    ReownTest,
 }
 
 #[derive(Error, Debug)]
@@ -107,6 +113,7 @@ impl ExchangeType {
         match self {
             ExchangeType::Binance => Box::new(BinanceExchange),
             ExchangeType::Coinbase => Box::new(CoinbaseExchange),
+            ExchangeType::ReownTest => Box::new(TestExchange),
         }
     }
 
@@ -126,6 +133,7 @@ impl ExchangeType {
         match self {
             ExchangeType::Binance => BinanceExchange.get_buy_url(state, params).await,
             ExchangeType::Coinbase => CoinbaseExchange.get_buy_url(state, params).await,
+            ExchangeType::ReownTest => TestExchange.get_buy_url(state, params),
         }
     }
 
@@ -137,6 +145,7 @@ impl ExchangeType {
         match self {
             ExchangeType::Binance => BinanceExchange.get_buy_status(state, params).await,
             ExchangeType::Coinbase => CoinbaseExchange.get_buy_status(state, params).await,
+            ExchangeType::ReownTest => TestExchange.get_buy_status(state, params).await,
         }
     }
 
@@ -167,6 +176,12 @@ pub fn is_feature_enabled_for_project_id(
     state: State<Arc<AppState>>,
     project_id: &String,
 ) -> Result<(), ExchangeError> {
+    if let Some(testing_project_id) = state.config.server.testing_project_id.as_ref() {
+        if crypto::constant_time_eq(testing_project_id, project_id) {
+            return Ok(());
+        }
+    }
+
     let allowed_project_ids = state
         .config
         .exchanges