+++ terraform/ecs/cluster.tf
@@ -84,7 +84,7 @@ resource "aws_ecs_task_definition" "app_task" {
         { name = "RPC_PROXY_GEOIP_DB_KEY", value = var.geoip_db_key },
         { name = "RPC_PROXY_TESTING_PROJECT_ID", value = var.testing_project_id },
 
-        { name = "RPC_PROXY_BLOCKED_COUNTRIES", value = var.ofac_blocked_countries },
+        { name = "RPC_PROXY_BLOCKED_COUNTRIES", value = var.ofac_countries },
 
         { name = "RPC_PROXY_PROVIDER_POKT_PROJECT_ID", value = var.pokt_project_id },
         { name = "RPC_PROXY_PROVIDER_QUICKNODE_API_TOKENS", value = var.quicknode_api_tokens },
@@ -106,6 +106,8 @@ resource "aws_ecs_task_definition" "app_task" {
         { name = "RPC_PROXY_PROVIDER_MELD_API_URL", value = var.meld_api_url },
         { name = "RPC_PROXY_PROVIDER_CALLSTATIC_API_KEY", value = var.callstatic_api_key },
         { name = "RPC_PROXY_PROVIDER_BLAST_API_KEY", value = var.blast_api_key },
+
+        { name = "RPC_PROXY_SKIP_QUOTA_CHAINS", value = var.proxy_skip_quota_chains },
 
         { name = "RPC_PROXY_PROVIDER_PROMETHEUS_ENDPOINT", value = var.prometheus_endpoint },
         { name = "RPC_PROXY_PROVIDER_PROMETHEUS_WORKSPACE_ID", value = var.prometheus_workspace_id },