+++ terraform/res_ecs.tf
@@ -59,7 +59,7 @@ module "ecs" {
   rate_limiting_cache_endpoint_read  = module.redis.endpoint
   rate_limiting_cache_endpoint_write = module.redis.endpoint
   provider_cache_endpoint            = module.redis.endpoint
-  ofac_blocked_countries             = var.ofac_blocked_countries
+  ofac_countries                     = var.ofac_countries
   postgres_url                       = module.postgres.database_url
 
   # Providers
@@ -83,6 +83,9 @@ module "ecs" {
   meld_api_url         = var.meld_api_url
   callstatic_api_key   = var.callstatic_api_key
   blast_api_key        = var.blast_api_key
+
+  # RPC Proxy configuration
+  proxy_skip_quota_chains = var.proxy_skip_quota_chains
 
   # Project Registry
   registry_api_endpoint   = var.registry_api_endpoint